king is now moving, added comments on what I did.
Need to added the same functions to the rest of the objects.

Make the P command work to change pc.

Implement the teleport.

In my opinion need to implement the teleport and end game and stuff to king and then add it to the rest of objects.

and so on so on so on so on    0_0